Delete ui_model/, dom_utils.ts, unused EJS templates, and remove initAppUI
