--- www/src/ui_model/debug.ts
@@ -1,161 +1,0 @@
-import { handleGitUserCaveatIssues } from "../github.js";
-import { ElementModel, ButtonElementModel } from "./model_lib.js";
-import { getFileManagerStore, FileKind, modifyDrawerNotificationForUserAddedFiles } from "../vue/file_manager_store.js";
-
-// This is in debug for now, since the only thing in it is the debug button
-// It is hidden by default, and only shown when something in it is enabled.
-class UserActionsGroup extends ElementModel {
-   public static readonly CLASS: string = "user-actions";
-
-   public static get(): UserActionsGroup {
-      return new UserActionsGroup(
-        ElementModel.getRequiredElementByQuery(`.${UserActionsGroup.CLASS}`));
-   }
-
-   /** @override */
-   public setHidden(hidden: boolean): void {
-      this.element.style.display = (hidden ? 'none' : 'block');
-  }
-  /** @override */
-  public isHidden(): boolean {
-      return this.element.style.display === 'none';
-  }
-}
-
-function isDebugModeEnabled(): boolean {
-   // Check if the hostname is localhost or some raw IP address.
-   const hostIsDebug = window.location.hostname === "localhost" ||
-      window.location.hostname.match(/^\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}$/);
-
-   // We can disable debug mode in the URL (mainly just for checking header layout)
-   const debugModeDisabled = window.location.search.includes("debug=false");
-
-   return (hostIsDebug && !debugModeDisabled) ? true : false;
-}
-
-export class DebugSettingsButton extends ButtonElementModel {
-   public static readonly ID: string = "debugSettingsButton";
-
-   public static get(): DebugSettingsButton {
-      return new DebugSettingsButton(
-         ElementModel.getRequiredElementById(DebugSettingsButton.ID));
-   }
-
-   public setup() {
-      this.setClickListener((_event) => {
-         new ElementModel(ElementModel.getRequiredElementById("debugSettingsPanel"))
-            .toggleHidden();
-      });
-
-      if (isDebugModeEnabled()) {
-         console.debug("Debug mode enabled");
-         UserActionsGroup.get().show();
-      }
-   }
-}
-
-class GithubOpenIssuesButton extends ButtonElementModel {
-   public static readonly ID: string = "generateGithubOpenIssuesButton";
-
-   public static get(): GithubOpenIssuesButton {
-      return new GithubOpenIssuesButton(
-         ElementModel.getRequiredElementById(GithubOpenIssuesButton.ID));
-   }
-
-   public setup() {
-      this.setClickListener((_event) => {
-         handleGitUserCaveatIssues(["some caveat"]);
-      });
-   }
-}
-
-class GithubFetchErrorButton extends ButtonElementModel {
-   public static readonly ID: string = "generateGithubFetchErrorButton";
-
-   public static get(): GithubFetchErrorButton {
-      return new GithubFetchErrorButton(
-         ElementModel.getRequiredElementById(GithubFetchErrorButton.ID));
-   }
-
-   public setup() {
-      this.setClickListener((_event) => {
-        const date = new Date();
-         if ((date.getSeconds() % 2) == 0) {
-            handleGitUserCaveatIssues({"message": "Sample issue from github"});
-         } else {
-            // Invalid format
-            handleGitUserCaveatIssues({"bla": "bar"});
-         }
-      });
-   }
-}
-
-class AddMockFilesButton extends ButtonElementModel {
-   public static readonly ID: string = "addMockFilesButton";
-
-   public static get(): AddMockFilesButton {
-      return new AddMockFilesButton(
-         ElementModel.getRequiredElementById(AddMockFilesButton.ID));
-   }
-
-   public setup() {
-      this.setClickListener((_event) => {
-         const store = getFileManagerStore();
-         const d = (k: FileKind) => FileKind.isDownloadableDefault(k);
-         store.addFile({ name: 'transactions_2024.csv',           kind: FileKind.AcbTxCsv,      isDownloadable: d(FileKind.AcbTxCsv),      useChecked: true,  data: new Uint8Array() });
-         store.addFile({ name: 'transactions_2023.csv',           kind: FileKind.AcbTxCsv,      isDownloadable: d(FileKind.AcbTxCsv),      useChecked: false, data: new Uint8Array() });
-         store.addFile({ name: 'acb_export_2024-01-15T12-00.zip', kind: FileKind.AcbOutputZip,  isDownloadable: d(FileKind.AcbOutputZip),  useChecked: false, data: new Uint8Array() });
-         store.addFile({ name: 'acb_export_2023-12-31T08-30.zip', kind: FileKind.AcbOutputZip,  isDownloadable: d(FileKind.AcbOutputZip),  useChecked: false, data: new Uint8Array() });
-         store.addFile({ name: 'summary_output.txt',              kind: FileKind.OutputText,    isDownloadable: d(FileKind.OutputText),    useChecked: false, data: new Uint8Array() });
-         store.addFile({ name: 'questrade_export.xlsx',           kind: FileKind.QuestradeXlsx, isDownloadable: d(FileKind.QuestradeXlsx), useChecked: true,  data: new Uint8Array() });
-         store.addFile({ name: 'unrecognized_file.dat',           kind: FileKind.Other,         isDownloadable: d(FileKind.Other),         useChecked: false, data: new Uint8Array(), warning: 'File type could not be determined' });
-         modifyDrawerNotificationForUserAddedFiles(store);
-      });
-   }
-}
-
-// When this checkbox is clicked, we'll re-load the page with debug_autoload=true
-// in the URL, which will trigger autoload when the page loads.
-export class AutoRunCheckbox extends ElementModel {
-   public static readonly ID: string = "debugAutoloadCheckbox";
-
-   public static get(): AutoRunCheckbox {
-      return new AutoRunCheckbox(
-         ElementModel.getRequiredElementById(AutoRunCheckbox.ID));
-   }
-
-   public checked(): boolean {
-      return (this.element as HTMLInputElement).checked;
-   }
-
-   public setup() {
-      const checked = isDebugModeEnabled() &&
-         window.location.search.includes("debug_autoload=true");
-      (this.element as HTMLInputElement).checked = checked;
-
-      this.element.addEventListener('change', () => {
-         // If checked, add debug_autoload=true to the URL, else
-         // remove it. In either case, reload the page with that URL.
-         const url = new URL(window.location.href);
-         if (AutoRunCheckbox.get().checked()) {
-            url.searchParams.set("debug_autoload", "true");
-         } else {
-            url.searchParams.delete("debug_autoload");
-         }
-         // Reload page
-         window.location.href = url.toString();
-      })
-   }
-}
-
-export class DebugSettings {
-   public static init() {
-      DebugSettingsButton.get().setup();
-
-      // Inside debug panel
-      GithubFetchErrorButton.get().setup();
-      GithubOpenIssuesButton.get().setup();
-      AddMockFilesButton.get().setup();
-      AutoRunCheckbox.get().setup();
-   }
-}
