--- www/templates/info_dialogs.ejs
@@ -1,96 +1,0 @@
-<div id="appDescriptionDialog" class="info-dialog">
-  <div class="info-dialog-header">
-    <h3 class="info-dialog-title">What is ACB?</h3>
-    <button class="info-dialog-close">&times;</button>
-  </div>
-  <div class="info-dialog-content">
-    <p><span class="tooltip">ACB<span class="tooltiptext">
-      Adjusted Cost Basis - The method in which capital gains tax are calculated in Canada. This can differ from the methods used in other countries.</span>
-      </span>
-      is a tool for tallying capital gains for Canadian tax returns.
-    </p>
-    <p>Features include:</p>
-    <ul class="info-list">
-      <li>Calculating <a href="https://www.canada.ca/en/revenue-agency/services/tax/individuals/topics/about-your-tax-return/tax-return/completing-a-tax-return/personal-income/line-127-capital-gains/capital-losses-deductions/what-a-superficial-loss.html">Superficial Losses</a></li>
-      <li>Automatic per-day foreign exchange rate lookups (USD only supported currently)</li>
-      <li>Can output to both table and text (good for saving for records)</li>
-    </ul>
-  </div>
-</div>
-
-<div id="fileFormatsDialog" class="info-dialog">
-  <div class="info-dialog-header">
-    <h3 class="info-dialog-title">Transactions CSV File Format</h3>
-    <button class="info-dialog-close">&times;</button>
-  </div>
-  <div class="info-dialog-content">
-    <p>Details on how to format your transactions CSV/spreadsheet can be found
-      <a href="https://github.com/tsiemens/acb/wiki/Transaction-Spreadsheets">here</a>.
-    </p>
-    <p><a href="./samples/sample_txs.csv">Sample transactions file</a>
-    </p>
-  </div>
-</div>
-
-<div id="liabilityDialog" class="info-dialog">
-  <div class="info-dialog-header">
-    <h3 class="info-dialog-title">Liability</h3>
-    <button class="info-dialog-close">&times;</button>
-  </div>
-  <div class="info-dialog-content">
-    <p>ACB is not designed or developed by certified individual, and is to be
-      used <strong>at your own risk</strong>. Please cross-check any results if you have any doubts.
-    </p>
-  </div>
-</div>
-
-<div id="dataPolicyDialog" class="info-dialog">
-  <div class="info-dialog-header">
-    <h3 class="info-dialog-title">Data Policy</h3>
-    <button class="info-dialog-close">&times;</button>
-  </div>
-  <div class="info-dialog-content">
-    <p>ACB does not collect any personal data. All processing is done locally to your browser.</p>
-  </div>
-</div>
-
-<div id="licenseDialog" class="info-dialog">
-  <div class="info-dialog-header">
-    <h3 class="info-dialog-title">License</h3>
-    <button class="info-dialog-close">&times;</button>
-  </div>
-  <div class="info-dialog-content">
-    <p>ACB is open sourced under the MIT License</p>
-    <div style="font-family: monospace;">
-      <p>
-        MIT License
-      </p>
-      <p>
-        Copyright (c) 2018 Trevor Siemens
-      </p>
-      <p>
-          Permission is hereby granted, free of charge, to any person obtaining a copy
-of this software and associated documentation files (the "Software"), to deal
-in the Software without restriction, including without limitation the rights
-to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-copies of the Software, and to permit persons to whom the Software is
-furnished to do so, subject to the following conditions:
-      </p>
-
-      <p>
-        The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-      </p>
-
-      <p>
-        THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-SOFTWARE.
-      </p>
-    </div>
-  </div>
-</div>
